Add crowdsourcing ignore list for persistent dissenters
Moderators can ignore players who consistently submit wrong crowdsourcing
values (e.g. due to a janky mafia setup or malicious intent). Ignored players
are silently excluded from all dissent Discord alerts.

- Adds crowdsourcingIgnored column to Player table (migration 008)
- Consensus alerts include an Ignore button per dissenter, mod-only
- Disagreeing submission alerts are suppressed for ignored players
- alert() now accepts string | MessageCreateOptions for richer messages
- Admin dailies UI shows ignored submissions greyed out with an Unignore button
- API: POST/DELETE /api/admin/dailies/ignore/:playerId
